Move some code around
