+++ chi-compiler/src/main/kotlin/gh/marad/chi/core/parser/ParserVisitor.kt
@@ -75,10 +75,10 @@ internal class ParserVisitor(private val source: ChiSource) : ChiParserBaseVisit
         )
 
     override fun visitFieldAccessExpr(ctx: ChiParser.FieldAccessExprContext): ParseAst =
-        DotOpReader.readFieldAccess(this, source, ctx)
+        FieldOperatorReader.readFieldAccess(this, source, ctx)
 
     override fun visitFieldAssignment(ctx: ChiParser.FieldAssignmentContext): ParseAst =
-        DotOpReader.readFieldAssignment(this, source, ctx)
+        FieldOperatorReader.readFieldAssignment(this, source, ctx)
 
 
     override fun visitWhileLoopExpr(ctx: ChiParser.WhileLoopExprContext): ParseAst =
