+++ chi-truffle-language/src/main/java/gh/marad/chi/truffle/Converter.java
@@ -13,10 +13,16 @@ import gh.marad.chi.truffle.nodes.ChiNode;
 import gh.marad.chi.truffle.nodes.FnRootNode;
 import gh.marad.chi.truffle.nodes.IndexOperatorNodeGen;
 import gh.marad.chi.truffle.nodes.IndexedAssignmentNodeGen;
-import gh.marad.chi.truffle.nodes.expr.*;
+import gh.marad.chi.truffle.nodes.expr.BlockExpr;
+import gh.marad.chi.truffle.nodes.expr.ExpressionNode;
+import gh.marad.chi.truffle.nodes.expr.IsNodeGen;
 import gh.marad.chi.truffle.nodes.expr.cast.CastToFloatNodeGen;
 import gh.marad.chi.truffle.nodes.expr.cast.CastToLongExprNodeGen;
 import gh.marad.chi.truffle.nodes.expr.cast.CastToStringNodeGen;
+import gh.marad.chi.truffle.nodes.expr.flow.IfExpr;
+import gh.marad.chi.truffle.nodes.expr.flow.WhileBreakNode;
+import gh.marad.chi.truffle.nodes.expr.flow.WhileContinueNode;
+import gh.marad.chi.truffle.nodes.expr.flow.WhileExprNode;
 import gh.marad.chi.truffle.nodes.expr.operators.arithmetic.*;
 import gh.marad.chi.truffle.nodes.expr.operators.bit.BitAndOperatorNodeGen;
 import gh.marad.chi.truffle.nodes.expr.operators.bit.BitOrOperatorNodeGen;
