+++ chi-truffle-language/src/main/java/gh/marad/chi/truffle/nodes/expr/flow/IfExpr.java
@@ -1,9 +1,10 @@
-package gh.marad.chi.truffle.nodes.expr;
+package gh.marad.chi.truffle.nodes.expr.flow;
 
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.UnexpectedResultException;
 import com.oracle.truffle.api.profiles.ConditionProfile;
 import gh.marad.chi.truffle.nodes.ChiNode;
+import gh.marad.chi.truffle.nodes.expr.ExpressionNode;
 import gh.marad.chi.truffle.nodes.expr.operators.bool.LogicNotOperator;
 import gh.marad.chi.truffle.runtime.TODO;
 import gh.marad.chi.truffle.runtime.Unit;
