+++ chi-truffle-language/src/main/java/gh/marad/chi/truffle/nodes/expr/flow/WhileContinueNode.java
@@ -1,6 +1,7 @@
-package gh.marad.chi.truffle.nodes.expr;
+package gh.marad.chi.truffle.nodes.expr.flow;
 
 import com.oracle.truffle.api.frame.VirtualFrame;
+import gh.marad.chi.truffle.nodes.expr.ExpressionNode;
 
 public class WhileContinueNode extends ExpressionNode {
     @Override
