+++ chi-truffle-language/src/main/java/gh/marad/chi/truffle/nodes/expr/flow/WhileExprNode.java
@@ -1,14 +1,16 @@
-package gh.marad.chi.truffle.nodes.expr;
+package gh.marad.chi.truffle.nodes.expr.flow;
 
 import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.nodes.LoopNode;
 import gh.marad.chi.truffle.nodes.ChiNode;
+import gh.marad.chi.truffle.nodes.expr.ExpressionNode;
 import gh.marad.chi.truffle.runtime.Unit;
 
 
 public class WhileExprNode extends ExpressionNode {
-    @Child private LoopNode loop;
+    @Child
+    private LoopNode loop;
 
     public WhileExprNode(ChiNode condition, ChiNode loop) {
         this.loop = Truffle.getRuntime().createLoopNode(new WhileRepeatingNode(condition, loop));
